Refactor uint8_t -> int

+++ src/hardware.h
@@ -9,20 +9,20 @@ namespace hardware {
 		const int VISION_PORT = 5;
 
 		namespace drive {
-			const uint8_t LEFT_FRONT = 1;
-			const uint8_t LEFT_BACK = 10;
-			const uint8_t RIGHT_FRONT = 2;
-			const uint8_t RIGHT_BACK = 4;
+			const int LEFT_FRONT = 1;
+			const int LEFT_BACK = 10;
+			const int RIGHT_FRONT = 2;
+			const int RIGHT_BACK = 4;
 		}
 
 		namespace transmission {
-			const uint8_t BOTTOM = 7;
-			const uint8_t TOP = 8;
+			const int BOTTOM = 7;
+			const int TOP = 8;
 		}
 
 		namespace intake {
-			const uint8_t LEFT = 19;
-			const uint8_t RIGHT = 20;
+			const int LEFT = 19;
+			const int RIGHT = 20;
 		}
 
 		/**
